WB8: statusPanel projection thinning and audio orchestrator
