+++ src/ui/statusPanel/feedback.ts
@@ -1,17 +1,23 @@
 import {
-	clearStatusPanelUserMessageLock,
-	pushStatusPanelUserMessageLock,
 	setStatusPanelCancelAllPending,
 	setStatusPanelConcurrencyText,
 	setStatusPanelCoverArtDataUrl,
-	setStatusPanelJobItems,
 	setStatusPanelIsProcessing,
+	setStatusPanelJobItems,
 	setStatusPanelProgressPercentage,
 	setStatusPanelStatusText,
 	setStatusPanelStepColor,
 	setStatusPanelStepText,
 } from './viewState.svelte';
 import type { JobListItem } from './viewTypes';
+
+export {
+	clearTransientStatusMessageLock,
+	pushTransientStatusMessage,
+	showError,
+	showInfo,
+	showSuccess,
+} from './viewState.svelte';
 
 export function updateProgressBar(percentage: number): void {
 	setStatusPanelProgressPercentage(percentage);
@@ -50,26 +56,6 @@ export function setCancelAllButtonPending(isPending: boolean): void {
 	setStatusPanelCancelAllPending(isPending);
 }
 
-export function showError(message: string): void {
-	updateStepText(`Error: ${message}`, 'var(--text-error, #ef4444)');
-}
-
-export function showSuccess(message: string): void {
-	updateStepText(message, 'var(--text-success, #10b981)');
-}
-
-export function showInfo(message: string): void {
-	updateStepText(message);
-}
-
 export function renderJobList(jobs: JobListItem[]): void {
 	setStatusPanelJobItems(jobs);
-}
-
-export function pushTransientStatusMessage(message: string, ttlMs?: number): void {
-	pushStatusPanelUserMessageLock(message, ttlMs);
-}
-
-export function clearTransientStatusMessageLock(): void {
-	clearStatusPanelUserMessageLock();
 }
